+++ lib/validators/any/optional/index.d.ts
@@ -1,2 +1,1 @@
-export * from './required';
 export * from './optional';