+++ lib/validators/any/optional/optional.d.ts
@@ -1,0 +1,2 @@
+import { Validator } from '../../../validation';
+export declare function optional<D, K extends keyof D>(validators: Validator<Required<D>, K>[]): Validator<D, K>[];