+++ lib/validators/any/required/required.d.ts
@@ -1,2 +1,2 @@
-import { ValidationResponse } from '../../../validation';
-export declare function required(value: any, key: string): ValidationResponse;
+import { ValidationResponse, Validator } from '../../../validation';
+export declare function required<D, K extends keyof D>(validators?: Validator<Required<D>, K>[]): ((value: Exclude<D[K], undefined | null | '' | false>, key: K, values: D) => Promise<ValidationResponse<K>>)[];