+++ lib/validators/string/maxLength/maxLength.d.ts
@@ -2,4 +2,4 @@ import { ValidationResponse } from '../../../validation';
 export interface MaxLengthValidationError {
     max: number;
 }
-export declare function maxLength(max: number): (value: string | undefined, key: string) => ValidationResponse<MaxLengthValidationError>;
+export declare function maxLength<K>(max: number): (value: string, key: K) => ValidationResponse<K, MaxLengthValidationError>;