+++ lib/validators/string/minLength/minLength.d.ts
@@ -2,4 +2,4 @@ import { ValidationResponse } from '../../../validation';
 export interface MinLengthValidationError {
     min: number;
 }
-export declare function minLength(min: number): (value: string | undefined, key: string) => ValidationResponse<MinLengthValidationError>;
+export declare function minLength<K>(min: number): (value: string, key: K) => ValidationResponse<K, MinLengthValidationError>;