+++ lib/validators/string/phone/phone.d.ts
@@ -1,2 +1,2 @@
 import { ValidationResponse } from '../../../validation';
-export declare function phone(value: string | undefined, key: string): ValidationResponse;
+export declare function phone<K>(value: string, key: K): ValidationResponse<K>;