+++ lib/validators/string/reg/reg.d.ts
@@ -2,4 +2,4 @@ import { ValidationResponse } from '../../../validation';
 export interface RegValidationData {
     regId?: any;
 }
-export declare function reg(reg: RegExp, regId?: any): (value: string | undefined, key: string) => ValidationResponse<RegValidationData>;
+export declare function reg(reg: RegExp, regId?: any): <K>(value: string, key: K) => ValidationResponse<K, RegValidationData>;