+++ src/validators/string/maxLength/maxLength.ts
@@ -4,10 +4,8 @@ export interface MaxLengthValidationError {
   max: number
 }
 
-export function maxLength (max: number) {
-  return function (value: string | undefined, key: string): ValidationResponse<MaxLengthValidationError> {
-    if (!value) return
-
+export function maxLength <K> (max: number) {
+  return function (value: string, key: K): ValidationResponse<K, MaxLengthValidationError> {
     if (value.length > max) {
       return {
         error: ValidationErrors.maxLength,